+++ lib/index.ts
@@ -1,2 +1,3 @@
 export * from './constants';
 export { cn } from './utils'; 
+export * from './celebration';